feat: add songs backend

+++ backend/packages/utils/s3Client.go
@@ -1,17 +1,20 @@
 package utils
 
 import (
+	"bytes"
 	"context"
+	"errors"
 	"strings"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
+	"github.com/aws/smithy-go"
 )
 
 var (
-	S3_BUCKET_NAME = "films-local-server"
+	S3_BUCKET_NAME = "songs-local-server"
 )
 
 func NewS3Client(ctx context.Context) (*s3.Client, error) {
@@ -39,24 +42,24 @@ func NewS3Uploader(ctx context.Context, fns ...func(*manager.Uploader)) (*manage
 	return manager.NewUploader(s3Client, fns...), nil
 }
 
-// func HeadObject(ctx context.Context, key string, s3Client *s3.Client) error {
-// 	var e error
-// 	var headObjectInput = &s3.HeadObjectInput{
-// 		Key:    &key,
-// 		Bucket: &S3_BUCKET_NAME,
-// 	}
+func HeadObject(ctx context.Context, key string, s3Client *s3.Client) error {
+	var e error
+	var headObjectInput = &s3.HeadObjectInput{
+		Key:    &key,
+		Bucket: &S3_BUCKET_NAME,
+	}
 
-// 	if _, e = s3Client.HeadObject(ctx, headObjectInput); e != nil {
-// 		var apiErr smithy.APIError
-// 		if errors.As(e, &apiErr); e != nil {
-// 			if apiErr.ErrorCode() == "NotFound" {
-// 				return ErrS3ObjectNotFound
-// 			}
-// 		}
-// 		return e
-// 	}
-// 	return nil
-// }
+	if _, e = s3Client.HeadObject(ctx, headObjectInput); e != nil {
+		var apiErr smithy.APIError
+		if errors.As(e, &apiErr); e != nil {
+			if apiErr.ErrorCode() == "NotFound" {
+				return ErrS3ObjectNotFound
+			}
+		}
+		return e
+	}
+	return nil
+}
 
 func DeleteObject(ctx context.Context, key string, s3Client *s3.Client) error {
 	var e error
@@ -65,23 +68,22 @@ func DeleteObject(ctx context.Context, key string, s3Client *s3.Client) error {
 		Bucket: &S3_BUCKET_NAME,
 	}
 
+	if e := HeadObject(ctx, key, s3Client); e != nil {
+		return e
+	}
+
 	if _, e = s3Client.DeleteObject(ctx, deleteObjectInput); e != nil {
 		return e
 	}
 
-	// if e := HeadObject(ctx, key, s3Client); e != nil {
-	// 	if e != ErrS3ObjectNotFound {
-	// 		return e
-	// 	}
-	// }
 	return nil
 }
 
-func UploadObject(ctx context.Context, key string, body string, storageClass types.StorageClass, uploader *manager.Uploader) (*manager.UploadOutput, error) {
+func UploadObject(ctx context.Context, key string, m *Multipart, storageClass types.StorageClass, uploader *manager.Uploader) (*manager.UploadOutput, error) {
 	var putObjectInput = &s3.PutObjectInput{
 		Bucket:       &S3_BUCKET_NAME,
 		Key:          &key,
-		Body:         strings.NewReader(body),
+		Body:         bytes.NewReader(m.Body),
 		StorageClass: types.StorageClassStandard,
 	}
